Add cron job script for cnx production db slim dump
acquire_data.yml in cnx-deploy can dump the production database for
developers, but most of the time we don't really need the data to be
up to date.  A weekly cron job to create a production database dump so
developers can download it for testing and development.

Some requirements for the server running the cron job:
 - prod-db-slim-dump should be in /etc/cron.weekly, owned and executable
   by root
 - the server it is on needs to have postgres installed, so the script
   can create a temporary database copied from the production database
 - the server needs a way of accessing the production database, either
   by `ssh` or `pg_dump`